Update 2020-05-06-bash-fish-zsh.md (sp)
I had read your write up and found it informative for the purpose I was curious as to, I am new to the world of what makes a computer tick and have just been reading what I come across. Your site had said to make or recommend any changes, additions or possible typo's to do so in here, I apologize if I have done this incorrectly, but I changed or corrected one typo, it was the spelling of "customization" on line 37. Thank you for the plugins and the information.